拡張集計のキャッシュ機能を追加
