+++ KmyKeiba/Models/Race/AnalysisTable/AnalysisTableSurface.cs
@@ -107,7 +107,7 @@ namespace KmyKeiba.Models.Race.AnalysisTable
       this.ProgressMax.Value = this.Rows.Count;
     }
 
-    public async Task AnalysisAsync(MyContext db, IReadOnlyList<RaceFinder> finders, IReadOnlyList<AnalysisTableWeight> weights, bool isCacheOnly, bool isBulk = false)
+    public async Task AnalysisAsync(MyContext db, IReadOnlyList<RaceFinder> finders, IReadOnlyList<AnalysisTableWeight> weights, bool isCacheOnly, bool isBulk = false, AggregateRaceFinder? aggregateFinder = null)
     {
       this.IsLoading.Value = true;
       this.ProgressMax.Value = this.Rows.Count;
@@ -115,7 +115,7 @@ namespace KmyKeiba.Models.Race.AnalysisTable
 
       foreach (var row in this.Rows.OrderBy(r => r.Data.Output == AnalysisTableRowOutputType.Binary ? 0 : 1))
       {
-        await row.LoadAsync(this.Race, finders, weights, isCacheOnly, isBulk);
+        await row.LoadAsync(this.Race, finders, weights, isCacheOnly, isBulk, aggregateFinder);
         this.Progress.Value++;
       }
