+++ KmyKeiba/Models/Script/ScriptRace.cs
@@ -2,6 +2,7 @@
 using KmyKeiba.Data.Wrappers;
 using KmyKeiba.Models.Analysis;
 using KmyKeiba.Models.Race;
+using KmyKeiba.Models.Race.Finder;
 using Microsoft.ClearScript;
 using System;
 using System.Collections.Generic;
@@ -142,7 +143,7 @@ namespace KmyKeiba.Models.Script
     [ScriptMember("findRaceHorsesAsync")]
     public async Task<string> FindRaceHorsesAsync(string keys, int count, int offset = 0)
     {
-      var result = await this._race.Finder.FindRaceHorsesAsync(keys, count, offset, withoutFutureRacesForce: true);
+      var result = await ((IRaceFinder)this._race.Finder).FindRaceHorsesAsync(keys, count, offset, withoutFutureRacesForce: true);
       return JsonSerializer.Serialize(
         result.Items.Select(s => new ScriptRaceHorse(string.Empty, s.Data)).Take(count).ToArray(), ScriptManager.JsonOptions);
     }
